backend/requirements.txt
Django==4.2.8
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework==3.14.0
markdown==3.5.1
django-filter==23.5
django-extensions==3.2.3
gunicorn==21.2.0
asgiref==3.7.2
packaging==23.2
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
pandas==2.1.4
openpyxl==3.1.2
django-cors-headers==4.3.1